requirements.txt
discord.py[voice]>=2.4.0
yt-dlp
python-dotenv>=1.0.0
PyNaCl>=1.5.0
